handling when user not logged in
