+++ js/src/components/interact/interact_v1.svelte
@@ -4,12 +4,12 @@
   import Nothing from "./routes/nothing.svelte";
   import TotalRecall from "./routes/total_recall_v1.svelte";
   import Slideshow from "./routes/slideshow_v1.svelte";
-  import SpacedRepetition from "./routes/spaced_repetition_v1.svelte";
+  import SpacedRepetitionAdd from "./routes/spaced_repetition_v1.svelte";
 
   const routes = {
     "/play/total_recall": TotalRecall,
     "/play/slideshow": Slideshow,
-    "/interact/spaced_repetition": SpacedRepetition,
+    "/interact/spaced_repetition/add": SpacedRepetitionAdd,
     // Catch-all, must be last
     "*": Nothing
   };
